Added parallel RAD reader, changed TagMap
Added parallel RAD reader which parses the prelude and file-level tag
values, and then performs the work of a parallel chunk reader.  This is
a slightly higher-level interface than the parallel chunk reader (though
that also remains).

Changed TagMap to *own* rather than borrow the underlying tag
descriptions.  The original TagMap behavior is available in a new type
called TagViewMap.  Right now there is a *ton* of duplication between
the implementations for TagMap and TagViewMap; figure out how to address
this in the future.