More support for u128 tags
Added support for U128 tag types and genericized the AlevinFryReadRecord
type over the type of the barcode.  For convenience there are type
aliases for 64 and 128 bit barcodes, and the original
AlevinFryReadRecord type is aliased to the new type with a 64-bit
barcode width.

Implemented a generic read_from to read into the appropriate barcode
type with proper conversion traits. This should be a zero-cost
abstraction.

Added both fallible and non-fallible versions of these functions.